requirements.txt
flask==3.0.2
flask-sqlalchemy==3.1.1
python-dotenv==1.0.0
werkzeug==3.0.1